// apps/web/src/lib/api/model/generation-options.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * edcurate-api
 * OpenAPI spec version: 0.1.0
 */

/**
 * Optional parameters for artifact generation.
 */
export interface GenerationOptions {
  quantity?: 'fewer' | 'standard' | null;
  difficulty?: 'easy' | 'medium' | 'hard' | null;
  instructions?: string | null;
  language?: string | null;
}